HW_17/requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
python-dotenv==0.21.0
Werkzeug==2.2.2
SQLAlchemy==1.4.44
